Changes to work with Java 1.1OBEX Type-ID FixOBEX large data fixL2CAP PSMs -> Hexsingle byte read() in RfCommStream could return -1 instead of 0xffNative changes for Windows

+++ de/avetana/bluetooth/l2cap/L2CAPConnectionNotifierImpl.java
@@ -60,7 +60,7 @@ public class L2CAPConnectionNotifierImpl extends ConnectionNotifier implements L
     String m_serviceName=(String)parsedURL.getParameter("name");
     m_serviceName=(m_serviceName==null?"Avetana Service":m_serviceName);
     myRecord=LocalServiceRecord.createSerialSvcRecord(new UUID(url.getLocalServiceUUID(), false), m_serviceName,
-        parsedURL.getAttrNumber().intValue(), parsedURL.getProtocol());
+        parsedURL.getAttrNumber(), parsedURL.getProtocol());
     if(myRecord==null) throw new Exception("Not a valid Service Record!!!!!");
   }
